+++ tests/test_integration.py
@@ -17,10 +17,9 @@ class TestIntegration():
     def test_can_get_inpadoc(self):
         pat_no = '8118221'
         app = USApplication.objects.get(patent_number=pat_no)
-        print(app.inpadoc.title)
-        assert app.inpadoc.title == 'Data storage and access systems'
+        print(app.inpadoc)
+        assert app.inpadoc[0].title == 'Data Storage and Access Systems'
 
     def test_can_get_us_app_from_inpadoc(self):
         app = Inpadoc.objects.get('US9231827B2')
-        print(app.us_application.patent_title)
         assert app.us_application.patent_title == 'FORMALIZING, DIFFUSING AND ENFORCING POLICY ADVISORIES AND MONITORING POLICY COMPLIANCE IN THE MANAGEMENT OF NETWORKS'