+++ CHANGELOG.rst
@@ -2,8 +2,15 @@
 Changelog
 =========
 
+0.1.4 (2018-10-24)
+------------------
+* Support for USPTO Applications, Assignments, and PTAB documents
+* Support for EPO Inpadoc and EPO Register
+* Initial Installation and Getting Started Documentation
+
 0.0.1 (2018-10-15)
 ------------------
 
 * First release on PyPI.
 * Support for EPO OPS (except Classification data)
+